requirement.txt
mysql-connector-python
flask-restful
